Moving the exception arg type check implementation to expr_common for the moment, adding to expr/base.py
