+++ pyomo/core/base/objective.py
@@ -22,7 +22,8 @@ from pyomo.common.modeling import NOTSET
 from pyomo.common.formatting import tabular_writer
 from pyomo.common.timing import ConstructionTimer
 
-from pyomo.core.expr.numvalue import value, _type_check_exception_arg
+from pyomo.core.expr.expr_common import _type_check_exception_arg
+from pyomo.core.expr.numvalue import value
 from pyomo.core.expr.template_expr import templatize_rule
 from pyomo.core.base.component import ActiveComponentData, ModelComponentFactory
 from pyomo.core.base.global_set import UnindexedComponent_index
