+++ pyomo/core/expr/numvalue.py
@@ -437,15 +437,3 @@ _pyomo_constant_types.add(NumericConstant)
 
 # We use as_numeric() so that the constant is also in the cache
 ZeroConstant = as_numeric(0)
-
-
-def _type_check_exception_arg(cls, exception):
-    if exception is NOTSET:
-        return True
-    elif type(exception) is not bool:
-        raise ValueError(
-            f"{cls.ctype.__name__} '{cls.name}' was called with a non-bool "
-            f"argument for 'exception': {exception}"
-        )
-    else:
-        return exception
